Halve embedding storage: float16 instead of float32, 512 dims instead of 1024
Two independent, stacking changes to the embedding/text_embedding BLOB
columns (47.2MB combined at the old format):

- Storage precision: float32 -> float16. Ollama.embed() now returns a
  Float16Array; bufToVec reads it back with a /2 byte-width divisor.
  Needs native Float16Array, which only landed in Node v24 - bumped the
  project's minimum Node version (package.json engines, .nvmrc) rather
  than hand-roll IEEE 754 half-precision conversion. Prototyped a
  hand-rolled version first and cross-checked it against Bun's native
  implementation (Bun already had this) over ~2000 values: found one
  real 1-ULP rounding mismatch, exactly the kind of subtle bug a native
  implementation avoids for free.

- Dimension: model-native (1024 for qwen3-embedding:0.6b) -> 512,
  configurable via the new ollama.embedDimensions. qwen3-embedding
  supports Matryoshka Representation Learning; verified live against
  the real Ollama instance before relying on it - both 1024 and 512-dim
  outputs are L2-normalized, same-text cosine between them is ~1.0
  (direction preserved), and two unrelated articles shift only slightly
  (0.3775 -> 0.3870). Defaults to null (model's native dimension) since
  not every embedding model supports this - opt-in per model.

Both changes are folded into syncEmbeddingSpace's version key
(embedModel::embedDimensions::f16) so upgrading always triggers a
rebuild via reembedMissing, even for an install whose model/dimensions
config didn't change.

better-sqlite3's native addon is ABI-tied to the Node version it was
built under; upgrading to Node 24 needs one rebuild
(npm rebuild better-sqlite3), documented in the README.

+++ test/api.test.js
@@ -4,7 +4,7 @@ import { tempDb, testConfig, startOllamaStub } from './helpers.js';
 import { createApp } from '../src/server.js';
 import { recomputeScores } from '../src/scoring.js';
 
-const vec = (...values) => Buffer.from(Float32Array.from(values).buffer);
+const vec = (...values) => Buffer.from(Float16Array.from(values).buffer);
 
 let db;
 let base;
